requirements.txt
Pillow==10.0.1
Pillow==10.3.0
Requests==2.32.2
streamlit==1.35.0
streamlit_lottie==0.0.5
